requirements.txt
alembic==0.9.5
amqp==2.2.2
billiard==3.5.0.3
blinker==1.4
celery==4.1.0
certifi==2017.11.5
chardet==3.0.4
click==6.7
elasticsearch==6.2.0
Flask==0.12.2
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.1
Flask-WTF==0.14.2
httplib2==0.11.0
idna==2.6
itsdangerous==0.24
jieba==0.39
Jinja2==2.9.6
kombu==4.1.0
Mako==1.0.7
MarkupSafe==1.0
meld3==1.0.2
msgpack-python==0.5.6
olefile==0.44
pbkdf2==1.3
Pillow==4.3.0
PyMySQL==0.7.11
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.3
qiniu==7.2.0
redis==2.10.6
requests==2.18.4
six==1.11.0
SQLAlchemy==1.1.14
urllib3==1.22
uWSGI==2.0.15
vine==1.1.4
Werkzeug==0.12.2
Whoosh==2.7.4
WTForms==2.1
